Restructure Java code.
